requirements.txt
asgiref==3.8.1
Django==5.0.7
django-admin-honeypot-updated-2021==1.2.0
django-cleanup==8.1.0
django-htmx==1.18.0
django-ipware==7.0.1
pillow==10.4.0
python-environ==0.4.54
python-ipware==3.0.0
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
